email_sender_backend/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
# asgiref==3.8.1
# Django==5.1.4
psycopg2-binary==2.9.10
python-decouple
# et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.3
# tzdata==2024.2
dj-database-url
dj-database-url
dj-database-url
